refactor pipeline for human processing variant calling for non human, and qc Stage can now create temp dir Added few unit tests
