+++ tests/test_pipelines/test_variant_calling.py
@@ -93,22 +93,22 @@ class TestGATKStage():
 
 
 class TestVariantCalling(TestAnalysisDriver):
-    dataset = NamedMock(
-        real_name='test_dataset',
-        user_sample_id='test_user_sample_id',
-        genome_version='genome_version',
-        reference_genome='reference_genome'
-    )
+
+    def setUp(self):
+        self.dataset = NamedMock(
+            real_name='test_dataset',
+            user_sample_id='test_user_sample_id',
+            genome_version='genome_version',
+            reference_genome='reference_genome'
+        )
 
 
 class TestBaseRecal(TestVariantCalling):
 
-    def setUp(self):
-        self.b = BaseRecal(dataset=self.dataset)
-
     def test_run(self):
+        b = BaseRecal(dataset=self.dataset)
         with patch_executor as e:
-            self.b._run()
+            b._run()
             assert e.call_count == 1
             e.assert_called_with("path/to/java_8 -Djava.io.tmpdir=tests/assets/jobs/test_dataset/gatk_var_calling "
                                  "-XX:+UseSerialGC "
@@ -132,12 +132,11 @@ class TestBaseRecal(TestVariantCalling):
 
 class TestPrintReads(TestVariantCalling):
 
-    def setUp(self):
-        self.p = PrintReads(dataset=self.dataset)
-
     def test_run(self):
+        p = PrintReads(dataset=self.dataset)
+
         with patch_executor as e:
-            self.p._run()
+            p._run()
             assert e.call_count == 1
             e.assert_called_with('path/to/java_8 -Djava.io.tmpdir=tests/assets/jobs/test_dataset/gatk_var_calling '
                                  '-XX:+UseSerialGC '
@@ -159,12 +158,12 @@ class TestPrintReads(TestVariantCalling):
 
 
 class TestRealignTarget(TestVariantCalling):
-    def setUp(self):
-        self.p = RealignTarget(dataset=self.dataset)
 
     def test_run(self):
+        p = RealignTarget(dataset=self.dataset)
+
         with patch_executor as e:
-            self.p._run()
+            p._run()
             assert e.call_count == 1
             e.assert_called_with('path/to/java_8 -Djava.io.tmpdir=tests/assets/jobs/test_dataset/gatk_var_calling '
                                  '-XX:+UseSerialGC '
@@ -185,12 +184,11 @@ class TestRealignTarget(TestVariantCalling):
 
 
 class TestRealign(TestVariantCalling):
-    def setUp(self):
-        self.p = Realign(dataset=self.dataset)
 
     def test_run(self):
+        p = Realign(dataset=self.dataset)
         with patch_executor as e:
-            self.p._run()
+            p._run()
             assert e.call_count == 1
             e.assert_called_with('path/to/java_8 -Djava.io.tmpdir=tests/assets/jobs/test_dataset/gatk_var_calling '
                                  '-XX:+UseSerialGC '
@@ -211,12 +209,12 @@ class TestRealign(TestVariantCalling):
 
 
 class TestHaplotypeCaller(TestVariantCalling):
-    def setUp(self):
-        self.p = HaplotypeCaller(dataset=self.dataset, input_bam='test_bam')
 
     def test_run(self):
+        p = HaplotypeCaller(dataset=self.dataset, input_bam='test_bam')
+
         with patch_executor as e:
-            self.p._run()
+            p._run()
             assert e.call_count == 3  # Command + bgzip + tabix
             assert e.call_args_list[0] == call(
                 'path/to/java_8 -Djava.io.tmpdir=tests/assets/jobs/test_dataset/gatk_var_calling '
@@ -260,12 +258,12 @@ class TestHaplotypeCaller(TestVariantCalling):
 
 
 class TestGenotypeGVCFs(TestVariantCalling):
-    def setUp(self):
-        self.p = GenotypeGVCFs(dataset=self.dataset)
 
     def test_run(self):
+        p = GenotypeGVCFs(dataset=self.dataset)
+
         with patch_executor as e:
-            self.p._run()
+            p._run()
             assert e.call_count == 3  # Command + bgzip + tabix
             assert e.call_args_list[0] == call(
                 'path/to/java_8 -Djava.io.tmpdir=tests/assets/jobs/test_dataset/gatk_var_calling '
@@ -289,12 +287,12 @@ class TestGenotypeGVCFs(TestVariantCalling):
 
 
 class TestSelectVariants(TestVariantCalling):
-    def setUp(self):
-        self.p = SelectVariants(dataset=self.dataset)
 
     def test_run(self):
+        p = SelectVariants(dataset=self.dataset)
+
         with patch_executor as e:
-            self.p._run()
+            p._run()
             assert e.call_count == 3  # Command + bgzip + tabix
             assert e.call_args_list[0] == call(
                 'path/to/java_8 -Djava.io.tmpdir=tests/assets/jobs/test_dataset/gatk_var_calling '
@@ -318,12 +316,12 @@ class TestSelectVariants(TestVariantCalling):
 
 
 class TestVariantFiltration(TestVariantCalling):
-    def setUp(self):
-        self.p = VariantFiltration(dataset=self.dataset)
 
     def test_run(self):
+        p = VariantFiltration(dataset=self.dataset)
+
         with patch_executor as e:
-            self.p._run()
+            p._run()
             assert e.call_count == 3  # Command + bgzip + tabix
             assert e.call_args_list[0] == call(
                 "path/to/java_8 -Djava.io.tmpdir=tests/assets/jobs/test_dataset/gatk_var_calling "
